complete v1

+++ application/views/page_admin/recive.php
@@ -35,16 +35,19 @@
             <div class="row">
                 <div class="col-lg-8 col-lg-offset-2" >
                     <div class="section-heading req-quote">
-                        <h2>dashboard </h2>
-                        <p>New quote to day <?= $date ?></p>
+                        <h2>Recive </h2>
+                        <!--
                         <form class="form-inline" action="/action_page.php">
+                        
                             <div class="form-group">
                                 <label for="search">search : </label>
                                 <input type="text"  class="form-control" name="search" id="search">
                             </div>
+
                             
                             <button type="submit" class="btn btn-default"><span class="glyphicon glyphicon-search"></span></button>
                         </form> 
+                        -->
                         <br>
                        
                         <div class="article" style="text-align: left;" >
@@ -55,35 +58,45 @@
                                     <th>company</th>
                                     <th>total + tex rate</th>
                                     <th>status</th>
-                                    <th colspan="2">option</th>
+                                    <th colspan="3">option</th>
                                 </tr>
                                 <?php
+                                if(empty($quotes) || $quotes[0]->quote_id ==null){
+                                    echo "<td colspan='7'>don't have quote</td>";
+                                }else{
                                     foreach ($quotes as $key => $value) {
                                         
                                         echo "<tr>";
                                         echo "<td><a href='".base_url()."index.php/page/gen_pdf/{$value->quote_no}'>{$value->quote_no}</a></td>";
-                                        echo "<td>{$value->customer_company}</td>";
-                                        echo "<td>{$value->total}</td>";
+                                        echo "<td>".htmlspecialchars($value->customer_company)."</td>";
+                                        echo "<td>".htmlspecialchars($value->total)."</td>";
                                         switch($value->quote_status){
                                             case 0:
                                                 echo "<td>not vertify</td>"; 
-                                                echo "<td><a href='http://www.google.com' onclick='return confirm(`Are you sure for change status to working?`)' >";
-                                                echo '<span class="glyphicon glyphicon-ok"></span>';
-                                                echo "</a></td>";
+                                                echo "<td>";
+                                                echo "<form action='".base_url()."index.php/page_admin/changeStatusQuote_quote_to_working' method='post'>";
+                                                echo "<input type='hidden' name='qid' value='{$value->quote_id}'>";
+                                                echo "<button type='submit' onclick='return confirm(`Are you sure for change status to working?`)'>"."<span class='glyphicon glyphicon-ok'></span>"."</button>";
+                                                
+                                                echo "</form>";
+                                                echo "</td>";
                                             break;
-                                            case 1: echo "<td>working</td>"; break;
-                                            case 2: echo "<td>complete</td>"; break;
+                                            case 1: echo "<td><a href='".base_url()."index.php/page_admin/working/{$value->quote_id}'>working</a></td>"; break;
+                                            case 2: echo "<td><a href='".base_url()."index.php/page_admin/recieve_detail/{$value->quote_id}'>receive</a></td>"; break;
                                         }
+                                        echo "<td><span class='glyphicon glyphicon-envelope'></span></td>";
+                                        //echo "<td><span class='glyphicon glyphicon-pencil'></span></td>";
+                                        echo "<td>
+                                                <form action='".base_url()."index.php/page_admin/delete_quote_recive' method='post'>
+                                                <input type='hidden' name='qid' value='{$value->quote_id}'>
+                                                <button  name='submit' type='subbmit' onclick='return confirm(`Are you want to delete {$value->quote_no}?`)'><span class='glyphicon glyphicon-trash'></span></button>
+                                                </form>
+                                            </td>";
                                         
-                                        
-                                        echo "<td><span class='glyphicon glyphicon-pencil'></span></td>";
-                                        echo "<td><span class='glyphicon glyphicon-trash'></span></td>";
-                                        
-
                                         echo "</tr>";
                                         
                                     }
-
+                                }
                                 ?>
 
 
@@ -105,9 +118,13 @@
             </div>
             </div>
         </div>
+        <ul class="pagination" id="pageing">
+    
+        </ul>
         </section>
-
-
+    
+   
+    
      <!-- Core JavaScript Files -->
      <script src="<?= base_url() ?>share/js/jquery.min.js"></script>
     <script src="<?= base_url() ?>share/js/bootstrap.min.js"></script>
@@ -119,6 +136,20 @@
 	<script src="<?= base_url() ?>share/js/stellar.js"></script>
     <!-- Custom Theme JavaScript -->
     <script src="<?= base_url() ?>share/js/custom.js"></script>
+    <script>
+        $(document).ready(()=>{
+            let html="";
+            for(i =1;i<= <?= $numpage ?>;i++){
+                if(<?= $page ?> ==  i){
+                    html = `<li class="disabled"><a href="#">${i}</a></li>`;
+                }else{
+                    html = `<li><a href="<?= base_url() ?>index.php/page_admin/recive/${i}">${i}</a></li>`;
+                }
+                $("#pageing").append(html);
+            }
 
+            
+        })
+    </script>
 </body>
 </html>
